pyproject.toml
[project]
name = "resurface"
version = "0.1.0"
description = "A lab for building and evaluating a Phase 0 heuristic classifier for screenshots"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.37.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "pytest>=7.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

